feat: VoucherFactory를 생성하여 Voucher 생성을 담당 refactor: VoucherType Enum class 간소화
