+++ src/test/java/com/programmers/voucher/voucher/VoucherTypeTest.java
@@ -3,7 +3,9 @@ package com.programmers.voucher.voucher;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.programmers.voucher.voucher.VoucherType.*;
+import static com.programmers.voucher.voucher.VoucherType.FixedAmount;
+import static com.programmers.voucher.voucher.VoucherType.PercentDiscount;
+import static com.programmers.voucher.voucher.VoucherValidator.getValidateVoucherType;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -27,28 +29,10 @@ class VoucherTypeTest {
     @DisplayName("VoucherType에 없는 타입을 검색한 경우 런타임 예외가 발생한다.")
     void 바우처타입검색() {
         assertThrows(RuntimeException.class,
-                () -> VoucherType.getValidateVoucherType("K"));
+                () -> getValidateVoucherType("K"));
     }
 
-    @Test
-    @DisplayName("정상적인 타입을 입력한 경우 VoucherType을 반환한다.")
-    void 바우처타입검색2() {
-        VoucherType fixedType = getValidateVoucherType("F");
-        assertThat(fixedType).isInstanceOf(VoucherType.class);
 
-        VoucherType percentType = getValidateVoucherType("P");
-        assertThat(percentType).isInstanceOf(VoucherType.class);
-    }
-
-    @Test
-    @DisplayName("정상적인 타입을 입력한 경우 VoucherType을 반환한다.")
-    void 바우처타입검색3() {
-        VoucherType fixedType = getValidateVoucherType("FixedAmount");
-        assertThat(fixedType).isInstanceOf(VoucherType.class);
-
-        VoucherType percentType = getValidateVoucherType("PercentDiscount");
-        assertThat(percentType).isInstanceOf(VoucherType.class);
-    }
 
 
 }
