+++ src/test/java/com/programmers/voucher/voucher/VoucherValidatorTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import static com.programmers.voucher.voucher.VoucherType.FixedAmount;
 import static com.programmers.voucher.voucher.VoucherType.PercentDiscount;
 import static com.programmers.voucher.voucher.VoucherValidator.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -100,5 +101,25 @@ class VoucherValidatorTest {
 
         assertThrows(RuntimeException.class,
                 () -> ValidateValue(PercentDiscount, String.valueOf(smallerThanMinValue)));
+    }
+
+    @Test
+    @DisplayName("정상적인 타입을 입력한 경우 VoucherType을 반환한다.")
+    void 바우처타입검색2() {
+        VoucherType fixedType = getValidateVoucherType("F");
+        assertThat(fixedType).isInstanceOf(VoucherType.class);
+
+        VoucherType percentType = getValidateVoucherType("P");
+        assertThat(percentType).isInstanceOf(VoucherType.class);
+    }
+
+    @Test
+    @DisplayName("정상적인 타입을 입력한 경우 VoucherType을 반환한다.")
+    void 바우처타입검색3() {
+        VoucherType fixedType = getValidateVoucherType("FixedAmount");
+        assertThat(fixedType).isInstanceOf(VoucherType.class);
+
+        VoucherType percentType = getValidateVoucherType("PercentDiscount");
+        assertThat(percentType).isInstanceOf(VoucherType.class);
     }
 }
